[step2] refactor: Lines 일급컬렉션 추가
