+++ src/main/java/ladder/domain/Ladder.java
@@ -7,7 +7,7 @@ import java.util.stream.IntStream;
 
 public class Ladder {
 
-    private final List<Line> lines;
+    private final Lines lines;
     private final int height;
 
     public Ladder(int lineSize, int height) {
@@ -16,28 +16,13 @@ public class Ladder {
             throw new IllegalArgumentException("라인 개수 또는 길이가 유효하지 않습니다.");
         }
 
-        this.lines = IntStream.range(0, lineSize)
-                .mapToObj(lineNumber -> createLine(lineSize, lineNumber, height))
-                .collect(Collectors.toList());
+        this.lines = Lines.of(lineSize, height);
         this.height = height;
-    }
-
-    private Line createLine(int lineSize, int lineNumber, int height) {
-
-        if (isLastLine(lineSize, lineNumber)) {
-            return Line.ofLastLine(height);
-        }
-        return Line.of(lineNumber, height);
-    }
-
-    private boolean isLastLine(int lineSize, int lineNumber) {
-
-        return lineNumber >= lineSize - 1;
     }
 
     public List<Line> getLines() {
 
-        return Collections.unmodifiableList(lines);
+        return lines.getLines();
     }
 
     public int getHeight() {
